requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
cryptography==41.0.3; python_version >= '3.7'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.5; python_version >= '3.8'
django-filter==23.3; python_version >= '3.7'
django-templated-mail==1.1.1
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.0; python_version >= '3.7'
djoser==2.2.0; python_version >= '3.8' and python_version < '4.0'
drf-yasg==1.21.7; python_version >= '3.6'
gunicorn==21.2.0; python_version >= '3.5'
idna==3.4; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
packaging==23.1; python_version >= '3.7'
pycparser==2.21
pyjwt==2.8.0; python_version >= '3.7'
python-decouple==3.8
python-dotenv==1.0.0; python_version >= '3.8'
python3-openid==3.2.0
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.3.0; python_version >= '3.7'
social-auth-core==4.4.2; python_version >= '3.6'
sqlparse==0.4.4; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.4; python_version >= '3.7'
whitenoise==6.5.0; python_version >= '3.7'
